Add files via upload
A join combines two tables into one result set. We can use joins when we want to query two tables at the same time. Whenever we join two tables, we have to specify the condition upon which the tables need to be joined.

In SQL, there are four different types of joins:

Inner Join
Left Outer Join
Right Outer Join
Full Outer Join
No matter which join you are working with, the general syntax for the query looks like so: